[refactor]: split VerifyPhonePage into two pages
